requirements.txt
psutil
numpy flask requests
mysql-connector-python
python-dotenv